chore: begins moving List view
