+++ packages/ui/src/elements/Paginator/ClickableArrow/index.scss
