+++ packages/ui/src/elements/Paginator/index.scss
