+++ packages/ui/src/views/List/SelectAll/index.scss
