+++ packages/ui/src/views/List/SelectAll/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { CheckboxInput } from '../../../../forms/field-types/Checkbox/Input'
 import { SelectAllStatus, useSelection } from '../SelectionProvider'
 import './index.scss'
+import { CheckboxInput } from '../../../forms/field-types/Checkbox/Input'
 
 const baseClass = 'select-all'
