+++ packages/ui/src/views/List/index.scss
@@ -1,4 +1,4 @@
-@import '../../../../scss/styles';
+@import '../../../scss/styles';
 
 .collection-list {
   width: 100%;
@@ -43,7 +43,9 @@
       width: 100%;
       overflow: auto;
 
-      [class^="cell"] > p, [class^="cell"] > span, [class^="cell"] > a {
+      [class^='cell'] > p,
+      [class^='cell'] > span,
+      [class^='cell'] > a {
         line-clamp: 4;
         -webkit-box-orient: vertical;
         -webkit-line-clamp: 4;
