+++ packages/ui/src/views/List/index.tsx
@@ -4,19 +4,18 @@ import { useTranslation } from 'react-i18next'
 import { useHistory } from 'react-router-dom'
 import { v4 as uuid } from 'uuid'
 
-import type { Where } from '../../../../../exports/types'
+import type { Where, Field } from 'payload/types'
 import type { ListIndexProps, ListPreferences, Props } from './types'
 
-import { type Field } from '../../../../../fields/config/types'
 import usePayloadAPI from '../../../../hooks/usePayloadAPI'
 import { useUseTitleField } from '../../../../hooks/useUseAsTitle'
 import { useStepNav } from '../../../elements/StepNav'
 import { TableColumnsProvider } from '../../../elements/TableColumns'
 import { useActions } from '../../../utilities/ActionsProvider'
-import { useAuth } from '../../../utilities/Auth'
-import { useConfig } from '../../../utilities/Config'
-import { usePreferences } from '../../../utilities/Preferences'
-import RenderCustomComponent from '../../../utilities/RenderCustomComponent'
+import { useAuth } from '../../providers/Auth'
+import { useConfig } from '../../providers/Config'
+import { usePreferences } from '../../providers/Preferences'
+import { RenderCustomComponent } from '../../elements/RenderCustomComponent'
 import { useSearchParams } from '../../../utilities/SearchParams'
 import DefaultList from './Default'
 import formatFields from './formatFields'
